@@ -1,9 +1,10 @@
 buildscript {
     repositories {
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.3'
+        classpath 'com.android.tools.build:gradle:3.0.0-alpha5'
     }
 }
 
@@ -34,13 +35,13 @@ task clean(type: Delete) {
 ext {
     minSdkVersion = 16
     targetSdkVersion = 25
-    compileSdkVersion = 25
-    buildToolsVersion = '25.0.3'
+    compileSdkVersion = 26
+    buildToolsVersion = '26.0.0'
 }
 
 ext.libs = [
-        supportVersion : '25.4.0',
-        constraintLayout : '1.0.2',
+        supportVersion : '26.0.0-beta2',
+        constraintLayout : '1.1.0-beta1',
 
         // for test
         junit : '4.12',
